Remove hardcoded float32s

+++ jcm/date.py
@@ -8,18 +8,18 @@ _DAYS_YEAR = 365.2425
 
 @tree_math.struct
 class DateData:
-    tyear: jnp.float32 # Fractional time of year, should possibly be part of the model itself (i.e. not in physics_data)
+    tyear: float # Fractional time of year, should possibly be part of the model itself (i.e. not in physics_data)
     model_year: jnp.int32
     model_step: jnp.int32
-    dt_seconds: jnp.float32 # Model timestep in seconds
+    dt_seconds: float # Model timestep in seconds
 
     @classmethod
     def zeros(cls, tyear=None, model_year=None, model_step=None, dt_seconds=None):
         return cls(
-          tyear=tyear if tyear is not None else jnp.float32(0.0),
+          tyear=tyear if tyear is not None else 0.0,
           model_year=model_year if model_year is not None else jnp.int32(1950),
           model_step=model_step if model_step is not None else jnp.int32(0),
-          dt_seconds=dt_seconds if dt_seconds is not None else jnp.float32(1800.0))
+          dt_seconds=dt_seconds if dt_seconds is not None else 1800.0)
 
     @classmethod
     def set_date(cls, model_time, model_step=None, dt_seconds=None):
@@ -27,15 +27,15 @@ class DateData:
           tyear=fraction_of_year_elapsed(model_time),
           model_year=get_year(model_time),
           model_step=model_step if model_step is not None else jnp.int32(0),
-          dt_seconds=dt_seconds if dt_seconds is not None else jnp.float32(1800.0))
+          dt_seconds=dt_seconds if dt_seconds is not None else 1800.0)
 
     @classmethod
     def ones(cls, tyear=None, model_year=None, model_step=None, dt_seconds=None):
         return cls(
-          tyear=tyear if tyear is not None else jnp.float32(1.0),
+          tyear=tyear if tyear is not None else 1.0,
           model_year=model_year if model_year is not None else jnp.int32(1950),
           model_step=model_step if model_step is not None else jnp.int32(0),
-          dt_seconds=dt_seconds if dt_seconds is not None else jnp.float32(1800.0))
+          dt_seconds=dt_seconds if dt_seconds is not None else 1800.0)
 
     def model_day(self):
         return jnp.round(self.tyear*_DAYS_YEAR).astype(jnp.int32)
@@ -73,4 +73,4 @@ def fraction_of_year_elapsed(dt: jdt.Datetime):
     days_elapsed_in_year += dt.delta.seconds / (24 * 60 * 60)
     
     # Calculate the fraction of the year elapsed
-    return jnp.float32(days_elapsed_in_year / _DAYS_YEAR)
+    return days_elapsed_in_year / _DAYS_YEAR
